refactor: move `isHovering` to `GameInterface`
